Added tool wc3trans for translating multiple maps
The tool takes a pair of war3map.wts files:
The original one and the translated version. It compares the translations
to a third one.
For every string which does also exist in the third one it uses the translated
string and generates a new automatically translated output file.

For the untranslated strings it just takes the old data.

Added a unit test as well.