[mpq] Fixed a bug with the wrong filename for sectors
Sectors had a filename with one directory before.
This led to errors on decrypting the sector data.
The filename without any directory has to be used on
decryption.

Besides use the uncompressed size for output buffers
when it is not equal to 0. This is usually smaller than
the archive's sector size.