[app] Add option --suggest to wc3trans
Don't show any warnings if the entry numbers are not the same since
only the first translation is used.

Add option --suggest which compares the untranslated string to all other
translated source strings if no translation exists and suggests to one
with the smallest difference.

This might help to find suggestions with minor differences and improve
the translation process.

Unfortunately, it also reduces the performance.